requirements.txt
gassist_text
google-auth
google-auth-oauthlib
google-assistant-sdk
paho-mqtt
cachetools
pytest
